feat: Move report tree into its own module [PT-188507981]
This is needed to clear out the Report module for when we create the reports context for the runs.
