+++ server/lib/report_server/application.ex
@@ -8,7 +8,7 @@ defmodule ReportServer.Application do
   @impl true
   def start(_type, _args) do
     # load the reports into an ets table
-    ReportServer.Reports.init()
+    ReportServer.Reports.Tree.init()
 
     children = [
       ReportServerWeb.Telemetry,
