+++ server/lib/report_server/reports/report.ex
@@ -1,0 +1,3 @@
+defmodule ReportServer.Reports.Report do
+  defstruct slug: nil, title: nil, subtitle: nil, run: nil, parents: [], path: nil
+end
