+++ server/lib/report_server_web/live/new_report_live/index.ex
@@ -2,11 +2,12 @@ defmodule ReportServerWeb.NewReportLive.Index do
   use ReportServerWeb, :live_view
 
   alias ReportServer.Reports
-  alias ReportServer.Reports.ReportGroup
+  alias ReportServer.Reports.Tree
+  alias ReportServer.Reports.Tree.ReportGroup
 
   @impl true
   def mount(%{"path" => path}, _session, socket) do
-    report_group = get_report_group(path)
+    report_group = Tree.find_report_group(path)
 
     socket = socket
       |> assign(:root_path, Reports.get_root_path())
@@ -14,15 +15,6 @@ defmodule ReportServerWeb.NewReportLive.Index do
       |> assign(:page_title, get_page_title(report_group))
 
     {:ok, socket}
-  end
-
-  defp get_report_group([]), do: Reports.tree()
-  defp get_report_group(path) do
-    slug = List.last(path)
-    case Reports.find(slug) do
-      report_group = %ReportGroup{} -> report_group
-      _ -> nil
-    end
   end
 
   defp get_page_title(nil), do: "Reports"
